Switched to floats from CGFloats to work better with AHEasing on 64 bit.

+++ UIView+EasingFunctions/UIView+EasingFunctions.h
@@ -8,7 +8,7 @@
 
 #import <UIKit/UIKit.h>
 
-typedef CGFloat (*ViewEasingFunctionPointerType)(CGFloat);
+typedef float (*ViewEasingFunctionPointerType)(float);
 
 /**
  Allows to assign custom [easing functions](http://easings.net/) to the animatable `UIView` properties.
